# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "智慧餐厅saas服务.",
        "title": "MP-ISV API",
        "contact": {
            "name": "codeCloud",
            "url": "https://codecloud.ltd",
            "email": "codeCloud2020@163.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4002",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/desk": {
            "get": {
                "description": "查看全部桌位列表",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 桌位管理"
                ],
                "summary": "桌位管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "桌位名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 获取成功，code:0 =\u003e 获取失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Paginate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Desk"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "提交删除单个桌位",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 桌位管理"
                ],
                "summary": "提交删除桌位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "单个菜单类型id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 删除成功，code:0 =\u003e 删除失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Desk"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            }
        },
...
        "/admin/desk/{id}": {
            "get": {
                "description": "查看单个桌位",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 桌位管理"
                ],
                "summary": "单个桌位详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "单个桌位id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 获取成功，code:0 =\u003e 获取失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Desk"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "提交修改单个桌位",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 桌位管理"
                ],
                "summary": "提交修改单个桌位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "单个桌位id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "单个桌位id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "小程序唯一编号",
                        "name": "mid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "桌位名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "所属门店",
                        "name": "store_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "所属分类",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 获取成功，code:0 =\u003e 获取失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Desk"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            },
            "delete": {
                "description": "提交删除桌位",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 桌位管理"
                ],
                "summary": "提交删除桌位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "单个桌位id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 删除成功，code:0 =\u003e 删除失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            }
        },
...
        "/admin/member": {
            "get": {
                "description": "查看全部会员列表",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "restaurant 会员管理"
                ],
                "summary": "会员管理",
                "responses": {
                    "200": {
                        "description": "code:1 =\u003e 获取成功，code:0 =\u003e 获取失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Paginate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"登录状态已失效！\"}",
                        "schema": {
                            "$ref": "#/definitions/gincmf_plugins_restaurantPlugin_model.ReturnData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gincmf_plugins_restaurantPlugin_model.ReturnData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.Desk": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "list_order": {
                    "type": "number"
                },
                "mid": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "store_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Paginate": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "birthday": {
                    "type": "integer"
                },
                "coin": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "integer"
                },
                "end_at": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "exp": {
                    "type": "integer"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_at": {
                    "type": "integer"
                },
                "last_loginip": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "more": {
                    "type": "string"
                },
                "open_id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "signature": {
                    "type": "string"
                },
                "start_at": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "update_at": {
                    "type": "integer"
                },
                "user_activation_key": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_login": {
                    "type": "string"
                },
                "user_nickname": {
                    "type": "string"
                },
                "user_realname": {
                    "type": "string"
                },
                "user_status": {
                    "type": "integer"
                },
                "user_type": {
                    "type": "integer"
                },
                "user_url": {
                    "type": "string"
                },
                "vip_can_open": {
                    "type": "boolean"
                },
                "vip_level": {
                    "type": "string"
                },
                "vip_name": {
                    "type": "string"
                },
                "vip_num": {
                    "type": "string"
                }
            }
        },
...
    }
}